Split comparison classes into one module each
